+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 - Improved performance when searching and browsing the storage overlay.
 - Made saving your storage data more robust if the game crashes or is force-closed.
+- Fixed pages with enchanted items sometimes failing to save.
 - Internal code cleanup and reliability improvements.
 - Maybe fixed a bug where the inventory items would go invisible when opening the storage page.